Gate ZSA issuance behind zsa-issuance feature
Address the second part of review comments r2542371427 - introduces zsa-issuance feature
flag to make issuance support optional while keeping the default build unchanged.

Based on pr471_various_fixes_1 branch (first part of r2542371427: move issuance-related
modules under crate::issuance::*), which is required for this change

- Add zsa-issuance feature (enabled by default) and wire it to an optional k256 dependency.
- Move issuance-related commitments code into bundle/commitments/issuance.rs and compile it
  only with zsa-issuance.
- Feature-gate issuance-dependent parts of note/asset_base.rs, including test-only stubs
  needed for run cargo tests without issuance.
- Gate issuance-only integration tests (and any issuance-only library tests/modules) behind
  zsa-issuance.
