+++ src/note/asset_base.rs
@@ -1,21 +1,28 @@
-use alloc::vec::Vec;
-use blake2b_simd::{Hash as Blake2bHash, Params};
 use core::cmp::Ordering;
 use core::hash::{Hash, Hasher};
-use group::{Group, GroupEncoding};
-use nonempty::NonEmpty;
+use group::GroupEncoding;
 use pasta_curves::{arithmetic::CurveExt, pallas};
-use rand_core::CryptoRngCore;
 use subtle::{Choice, ConstantTimeEq, CtOption};
 
-use crate::{
-    constants::fixed_bases::{
-        NATIVE_ASSET_BASE_V_BYTES, VALUE_COMMITMENT_PERSONALIZATION, ZSA_ASSET_BASE_PERSONALIZATION,
-    },
-    issuance::{
+use crate::constants::fixed_bases::{NATIVE_ASSET_BASE_V_BYTES, VALUE_COMMITMENT_PERSONALIZATION};
+
+// Conditional imports for ZSA issuance and/or tests
+#[cfg(any(feature = "zsa-issuance", test))]
+use {
+    crate::constants::fixed_bases::ZSA_ASSET_BASE_PERSONALIZATION, group::Group,
+    rand_core::CryptoRngCore,
+};
+
+// Conditional imports for full ZSA issuance feature
+#[cfg(feature = "zsa-issuance")]
+use {
+    crate::issuance::{
         auth::{IssueAuthKey, IssueValidatingKey, ZSASchnorr},
         compute_asset_desc_hash,
     },
+    alloc::vec::Vec,
+    blake2b_simd::{Hash as Blake2bHash, Params},
+    nonempty::NonEmpty,
 };
 
 /// Note type identifier.
@@ -36,6 +43,7 @@ impl Ord for AssetBase {
 }
 
 /// Personalization for the ZSA asset digest generator
+#[cfg(feature = "zsa-issuance")]
 pub const ZSA_ASSET_DIGEST_PERSONALIZATION: &[u8; 16] = b"ZSA-Asset-Digest";
 
 /// Derives the Asset Digest for the given ZSA asset.
@@ -43,6 +51,7 @@ pub const ZSA_ASSET_DIGEST_PERSONALIZATION: &[u8; 16] = b"ZSA-Asset-Digest";
 /// Defined in [ZIP-227: Issuance of Zcash Shielded Assets][assetdigest].
 ///
 /// [assetdigest]: https://zips.z.cash/zip-0227#asset-digests
+#[cfg(feature = "zsa-issuance")]
 pub fn asset_digest(encode_asset_id: &[u8]) -> Blake2bHash {
     Params::new()
         .hash_length(64)
@@ -55,6 +64,7 @@ pub fn asset_digest(encode_asset_id: &[u8]) -> Blake2bHash {
 /// Encoding the Asset Identifier, as defined in [ZIP 227][assetidentifier].
 ///
 /// [assetidentifier]: https://zips.z.cash/zip-0227.html#specification-asset-identifier-asset-digest-and-asset-base
+#[cfg(feature = "zsa-issuance")]
 pub fn encode_asset_id(
     version: u8,
     ik: &IssueValidatingKey<ZSASchnorr>,
@@ -88,6 +98,7 @@ impl AssetBase {
     /// # Panics
     ///
     /// Panics if the derived AssetBase is the identity point.
+    #[cfg(feature = "zsa-issuance")]
     #[allow(non_snake_case)]
     pub fn derive(ik: &IssueValidatingKey<ZSASchnorr>, asset_desc_hash: &[u8; 32]) -> Self {
         let version_byte: u8 = 0x00;
@@ -106,6 +117,22 @@ impl AssetBase {
         );
 
         // AssetBase = ZSAValueBase(AssetDigest)
+        AssetBase(asset_base)
+    }
+
+    /// Stub for derive when zsa-issuance is disabled.
+    /// Creates a deterministic asset base from a hash for testing without issuance keys.
+    #[cfg(all(test, not(feature = "zsa-issuance")))]
+    pub(crate) fn derive_stub(asset_desc_hash: &[u8; 32]) -> Self {
+        let asset_base =
+            pallas::Point::hash_to_curve(ZSA_ASSET_BASE_PERSONALIZATION)(asset_desc_hash);
+
+        // this will happen with negligible probability.
+        assert!(
+            bool::from(!asset_base.is_identity()),
+            "The Asset Base is the identity point, which is invalid."
+        );
+
         AssetBase(asset_base)
     }
 
@@ -129,6 +156,7 @@ impl AssetBase {
     /// Generates a ZSA random asset.
     ///
     /// This is only used in tests.
+    #[cfg(feature = "zsa-issuance")]
     pub(crate) fn random(rng: &mut impl CryptoRngCore) -> Self {
         let isk = IssueAuthKey::<ZSASchnorr>::random(rng);
         let ik = IssueValidatingKey::from(&isk);
@@ -136,6 +164,13 @@ impl AssetBase {
             &ik,
             &compute_asset_desc_hash(&NonEmpty::from_slice(b"zsa_asset").unwrap()),
         )
+    }
+
+    /// Stub for random when zsa-issuance is disabled.
+    /// Generates a random non-native asset for testing without issuance keys.
+    #[cfg(all(test, not(feature = "zsa-issuance")))]
+    pub(crate) fn random(rng: &mut impl CryptoRngCore) -> Self {
+        Self(pallas::Point::random(rng))
     }
 }
 
@@ -153,7 +188,7 @@ impl PartialEq for AssetBase {
 }
 
 /// Generators for property testing.
-#[cfg(any(test, feature = "test-dependencies"))]
+#[cfg(all(any(test, feature = "test-dependencies"), feature = "zsa-issuance"))]
 #[cfg_attr(docsrs, doc(cfg(feature = "test-dependencies")))]
 pub mod testing {
     use super::AssetBase;
@@ -199,7 +234,38 @@ pub mod testing {
     }
 }
 
+/// Generators for property testing (without zsa-issuance feature).
+#[cfg(all(
+    any(test, feature = "test-dependencies"),
+    not(feature = "zsa-issuance")
+))]
+#[cfg_attr(docsrs, doc(cfg(feature = "test-dependencies")))]
+pub mod testing {
+    use super::AssetBase;
+    use proptest::prelude::*;
+
+    prop_compose! {
+        /// Generate a uniformly distributed note type (native only without zsa-issuance feature)
+        pub fn arb_asset_base()
+            (_is_native in prop::bool::ANY)
+            -> AssetBase
+        {
+            AssetBase::native()
+        }
+    }
+
+    prop_compose! {
+        /// Generate a ZSA asset (stub without issuance keys)
+        pub fn arb_zsa_asset_base()(
+            asset_desc_hash in any::<[u8; 32]>(),
+        ) -> AssetBase {
+            AssetBase::derive_stub(&asset_desc_hash)
+        }
+    }
+}
+
 #[cfg(test)]
+#[cfg(feature = "zsa-issuance")]
 mod tests {
     use crate::{
         issuance::auth::{IssueValidatingKey, ZSASchnorr},
